requirements.txt
pynacl
discord.py[voice]
rethinkdb